+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ### Version 2.1.1
 - FIX: A bug preventing allergens create
+- FIX: Allergen import from PV Online
 
 ### Version 2.0.9
 - ADD: Purity percentage in cart